requirements.txt
streamlit==1.36.0
tensorflow-cpu==2.15.0
keras==2.15.0
numpy==1.26.4
pillow
pandas
scikit-learn
matplotlib

